backend/requirement.txt
# LangChain Core
langchain
langchain-core
langchain-community
langchain-text-splitters

# OpenAI Integration
langchain-openai
openai



# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn


sentence-transformers